# openapi.json
{"openapi":"3.1.0","info":{"title":"Pulse Plan API","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["auth"],"summary":"Signup","operationId":"signup_api_v1_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Get Current User Profile","operationId":"get_current_user_profile_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"Bearer":[]}]}},"/api/v1/auth/logout":{"post":{"tags":["auth"],"summary":"Logout","operationId":"logout_api_v1_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Logout Api V1 Auth Logout Post"}}}}}}},"/api/v1/preferences/me":{"get":{"tags":["preferences"],"summary":"Get Preferences","operationId":"get_preferences_api_v1_preferences_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreferencesResponse"}}}}}},"put":{"tags":["preferences"],"summary":"Upsert Preferences","operationId":"upsert_preferences_api_v1_preferences_me_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPreferencesUpsert"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreferencesMutationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/routine-tasks":{"get":{"tags":["preferences"],"summary":"List Routine Tasks","operationId":"list_routine_tasks_api_v1_routine_tasks_get","parameters":[{"name":"active_only","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Active Only"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutineTaskListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["preferences"],"summary":"Create Routine Task","operationId":"create_routine_task_api_v1_routine_tasks_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutineTaskCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutineTaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/routine-tasks/{routine_task_id}":{"put":{"tags":["preferences"],"summary":"Update Routine Task","operationId":"update_routine_task_api_v1_routine_tasks__routine_task_id__put","parameters":[{"name":"routine_task_id","in":"path","required":true,"schema":{"type":"string","title":"Routine Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutineTaskUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutineTaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["preferences"],"summary":"Delete Routine Task","operationId":"delete_routine_task_api_v1_routine_tasks__routine_task_id__delete","parameters":[{"name":"routine_task_id","in":"path","required":true,"schema":{"type":"string","title":"Routine Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutineTaskMessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks":{"get":{"tags":["tasks"],"summary":"List Tasks","operationId":"list_tasks_api_v1_tasks_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/TaskStatus"},{"type":"null"}],"title":"Status"}},{"name":"priority","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/Priority"},{"type":"null"}],"title":"Priority"}},{"name":"sort","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/TaskSort"},{"type":"null"}],"default":"deadline","title":"Sort"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["tasks"],"summary":"Create Task","operationId":"create_task_api_v1_tasks_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskMutationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/{task_id}":{"get":{"tags":["tasks"],"summary":"Get Task","operationId":"get_task_api_v1_tasks__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["tasks"],"summary":"Update Task","operationId":"update_task_api_v1_tasks__task_id__put","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskMessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["tasks"],"summary":"Delete Task","operationId":"delete_task_api_v1_tasks__task_id__delete","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskMessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/{task_id}/complete":{"patch":{"tags":["tasks"],"summary":"Complete Task","operationId":"complete_task_api_v1_tasks__task_id__complete_patch","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskMessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/schedule/generate":{"post":{"tags":["schedule"],"summary":"Generate Schedule","operationId":"generate_schedule_api_v1_schedule_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleGenerateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleOutput"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/schedule":{"get":{"tags":["schedule"],"summary":"Get Current Schedule","operationId":"get_current_schedule_api_v1_schedule_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CurrentScheduleResponse"}}}}}}},"/api/v1/schedule/reschedule":{"post":{"tags":["schedule"],"summary":"Reschedule","operationId":"reschedule_api_v1_schedule_reschedule_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleRescheduleRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatedScheduleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/orda/process":{"post":{"tags":["orda"],"summary":"Process Command","description":"Process natural language commands via AI.","operationId":"process_command_api_v1_orda_process_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ORDARequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrdaResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/orda/history":{"get":{"tags":["orda"],"summary":"Get History","description":"Retrieve ORDA conversation history.","operationId":"get_history_api_v1_orda_history_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrdaHistoryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AuthRequest":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":256,"minLength":8,"title":"Password"}},"type":"object","required":["email","password"],"title":"AuthRequest"},"CurrentScheduleResponse":{"properties":{"schedule":{"items":{"$ref":"#/components/schemas/ScheduleItem"},"type":"array","title":"Schedule"},"unscheduled_tasks":{"items":{"$ref":"#/components/schemas/UnscheduledTask"},"type":"array","title":"Unscheduled Tasks"}},"type":"object","required":["schedule"],"title":"CurrentScheduleResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Refresh Token"},"user":{"$ref":"#/components/schemas/User"}},"type":"object","required":["access_token","user"],"title":"LoginResponse"},"ORDARequest":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"ORDARequest"},"OrdaHistoryItem":{"properties":{"id":{"type":"string","title":"Id"},"message":{"type":"string","title":"Message"},"response":{"type":"string","title":"Response"},"timestamp":{"type":"string","title":"Timestamp"}},"type":"object","required":["id","message","response","timestamp"],"title":"OrdaHistoryItem"},"OrdaHistoryResponse":{"properties":{"history":{"items":{"$ref":"#/components/schemas/OrdaHistoryItem"},"type":"array","title":"History"}},"type":"object","required":["history"],"title":"OrdaHistoryResponse"},"OrdaResponse":{"properties":{"intent":{"type":"string","title":"Intent"},"response":{"type":"string","title":"Response"},"action_taken":{"type":"boolean","title":"Action Taken"},"data":{"type":"object","title":"Data"}},"type":"object","required":["intent","response","action_taken","data"],"title":"OrdaResponse"},"PreferencesMutationResponse":{"properties":{"message":{"type":"string","title":"Message"},"preferences":{"$ref":"#/components/schemas/UserPreferences"},"routine_tasks":{"items":{"$ref":"#/components/schemas/RoutineTask"},"type":"array","title":"Routine Tasks"}},"type":"object","required":["message","preferences"],"title":"PreferencesMutationResponse"},"PreferencesResponse":{"properties":{"preferences":{"anyOf":[{"$ref":"#/components/schemas/UserPreferences"},{"type":"null"}]},"routine_tasks":{"items":{"$ref":"#/components/schemas/RoutineTask"},"type":"array","title":"Routine Tasks"}},"type":"object","required":["preferences"],"title":"PreferencesResponse"},"Priority":{"type":"string","enum":["low","medium","high"],"title":"Priority"},"RoutineTask":{"properties":{"title":{"type":"string","maxLength":160,"minLength":1,"title":"Title"},"start_time":{"type":"string","format":"time","title":"Start Time"},"end_time":{"type":"string","format":"time","title":"End Time"},"days":{"items":{"type":"string"},"type":"array","title":"Days"},"is_active":{"type":"boolean","title":"Is Active","default":true},"id":{"type":"string","format":"uuid","title":"Id"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["title","start_time","end_time","id","user_id","created_at"],"title":"RoutineTask"},"RoutineTaskCreate":{"properties":{"title":{"type":"string","maxLength":160,"minLength":1,"title":"Title"},"start_time":{"type":"string","format":"time","title":"Start Time"},"end_time":{"type":"string","format":"time","title":"End Time"},"days":{"items":{"type":"string"},"type":"array","title":"Days"},"is_active":{"type":"boolean","title":"Is Active","default":true}},"type":"object","required":["title","start_time","end_time"],"title":"RoutineTaskCreate"},"RoutineTaskListResponse":{"properties":{"routine_tasks":{"items":{"$ref":"#/components/schemas/RoutineTask"},"type":"array","title":"Routine Tasks"}},"type":"object","required":["routine_tasks"],"title":"RoutineTaskListResponse"},"RoutineTaskMessageResponse":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"RoutineTaskMessageResponse"},"RoutineTaskResponse":{"properties":{"message":{"type":"string","title":"Message"},"routine_task":{"$ref":"#/components/schemas/RoutineTask"}},"type":"object","required":["message","routine_task"],"title":"RoutineTaskResponse"},"RoutineTaskUpdate":{"properties":{"title":{"anyOf":[{"type":"string","maxLength":160,"minLength":1},{"type":"null"}],"title":"Title"},"start_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Start Time"},"end_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"End Time"},"days":{"items":{"type":"string"},"type":"array","title":"Days"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"}},"type":"object","title":"RoutineTaskUpdate"},"ScheduleEvent":{"properties":{"title":{"type":"string","maxLength":160,"minLength":1,"title":"Title"},"start_time":{"type":"string","format":"date-time","title":"Start Time"},"end_time":{"type":"string","format":"date-time","title":"End Time"}},"type":"object","required":["title","start_time","end_time"],"title":"ScheduleEvent"},"ScheduleGenerateRequest":{"properties":{"available_hours":{"items":{"$ref":"#/components/schemas/TimeBlock"},"type":"array","minItems":1,"title":"Available Hours"},"productivity_preference":{"type":"string","maxLength":40,"title":"Productivity Preference"}},"type":"object","required":["available_hours","productivity_preference"],"title":"ScheduleGenerateRequest"},"ScheduleItem":{"properties":{"task_id":{"type":"string","title":"Task Id"},"task_title":{"type":"string","title":"Task Title"},"start_time":{"type":"string","format":"date-time","title":"Start Time"},"end_time":{"type":"string","format":"date-time","title":"End Time"},"part":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Part"},"total_parts":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Parts"}},"type":"object","required":["task_id","task_title","start_time","end_time"],"title":"ScheduleItem"},"ScheduleOutput":{"properties":{"schedule":{"items":{"$ref":"#/components/schemas/ScheduleItem"},"type":"array","title":"Schedule"},"unscheduled_tasks":{"items":{"$ref":"#/components/schemas/UnscheduledTask"},"type":"array","title":"Unscheduled Tasks"},"justification":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Justification"}},"type":"object","required":["schedule"],"title":"ScheduleOutput"},"ScheduleRescheduleRequest":{"properties":{"reason":{"type":"string","maxLength":500,"minLength":1,"title":"Reason"},"event":{"$ref":"#/components/schemas/ScheduleEvent"}},"type":"object","required":["reason","event"],"title":"ScheduleRescheduleRequest"},"SignupResponse":{"properties":{"message":{"type":"string","title":"Message","default":"User registered successfully"}},"type":"object","title":"SignupResponse"},"Task":{"properties":{"title":{"type":"string","maxLength":160,"minLength":1,"title":"Title"},"description":{"type":"string","title":"Description","default":""},"deadline":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deadline"},"estimated_duration_minutes":{"type":"integer","maximum":1440.0,"minimum":5.0,"title":"Estimated Duration Minutes"},"priority":{"$ref":"#/components/schemas/Priority","default":"medium"},"status":{"$ref":"#/components/schemas/TaskStatus","default":"pending"},"is_routine":{"type":"boolean","title":"Is Routine","default":false},"recurrence_rule":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Recurrence Rule"},"fixed_start_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Fixed Start Time"},"fixed_end_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Fixed End Time"},"id":{"type":"string","format":"uuid","title":"Id"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["title","estimated_duration_minutes","id","user_id","created_at","updated_at"],"title":"Task"},"TaskCreate":{"properties":{"title":{"type":"string","maxLength":160,"minLength":1,"title":"Title"},"description":{"type":"string","title":"Description","default":""},"deadline":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deadline"},"estimated_duration_minutes":{"type":"integer","maximum":1440.0,"minimum":5.0,"title":"Estimated Duration Minutes"},"priority":{"$ref":"#/components/schemas/Priority","default":"medium"},"status":{"$ref":"#/components/schemas/TaskStatus","default":"pending"},"is_routine":{"type":"boolean","title":"Is Routine","default":false},"recurrence_rule":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Recurrence Rule"},"fixed_start_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Fixed Start Time"},"fixed_end_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Fixed End Time"}},"type":"object","required":["title","estimated_duration_minutes"],"title":"TaskCreate"},"TaskListResponse":{"properties":{"tasks":{"items":{"$ref":"#/components/schemas/Task"},"type":"array","title":"Tasks"}},"type":"object","required":["tasks"],"title":"TaskListResponse"},"TaskMessageResponse":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"TaskMessageResponse"},"TaskMutationResponse":{"properties":{"message":{"type":"string","title":"Message"},"task":{"anyOf":[{"$ref":"#/components/schemas/Task"},{"type":"null"}]}},"type":"object","required":["message"],"title":"TaskMutationResponse"},"TaskResponse":{"properties":{"task":{"$ref":"#/components/schemas/Task"}},"type":"object","required":["task"],"title":"TaskResponse"},"TaskSort":{"type":"string","enum":["deadline","priority"],"title":"TaskSort"},"TaskStatus":{"type":"string","enum":["pending","scheduled","completed"],"title":"TaskStatus"},"TaskUpdate":{"properties":{"title":{"anyOf":[{"type":"string","maxLength":160,"minLength":1},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"deadline":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deadline"},"estimated_duration_minutes":{"anyOf":[{"type":"integer","maximum":1440.0,"minimum":5.0},{"type":"null"}],"title":"Estimated Duration Minutes"},"priority":{"anyOf":[{"$ref":"#/components/schemas/Priority"},{"type":"null"}]},"status":{"anyOf":[{"$ref":"#/components/schemas/TaskStatus"},{"type":"null"}]},"is_routine":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Routine"},"recurrence_rule":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Recurrence Rule"},"fixed_start_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Fixed Start Time"},"fixed_end_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Fixed End Time"}},"type":"object","title":"TaskUpdate"},"TimeBlock":{"properties":{"start":{"type":"string","format":"time","title":"Start"},"end":{"type":"string","format":"time","title":"End"}},"type":"object","required":["start","end"],"title":"TimeBlock"},"UnscheduledTask":{"properties":{"task_id":{"type":"string","title":"Task Id"},"task_title":{"type":"string","title":"Task Title"},"reason":{"type":"string","title":"Reason"}},"type":"object","required":["task_id","task_title","reason"],"title":"UnscheduledTask"},"UpdatedScheduleResponse":{"properties":{"updated_schedule":{"items":{"$ref":"#/components/schemas/ScheduleItem"},"type":"array","title":"Updated Schedule"},"unscheduled_tasks":{"items":{"$ref":"#/components/schemas/UnscheduledTask"},"type":"array","title":"Unscheduled Tasks"}},"type":"object","required":["updated_schedule"],"title":"UpdatedScheduleResponse"},"User":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name","default":""},"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["id","email"],"title":"User"},"UserPreferences":{"properties":{"productivity_period":{"anyOf":[{"type":"string","maxLength":40},{"type":"null"}],"title":"Productivity Period"},"work_style":{"anyOf":[{"type":"string","maxLength":80},{"type":"null"}],"title":"Work Style"},"focus_duration_minutes":{"anyOf":[{"type":"integer","maximum":180.0,"minimum":15.0},{"type":"null"}],"title":"Focus Duration Minutes"},"break_preference":{"anyOf":[{"type":"string","maxLength":80},{"type":"null"}],"title":"Break Preference"},"sleep_hours":{"anyOf":[{"type":"string","maxLength":80},{"type":"null"}],"title":"Sleep Hours"},"task_count":{"anyOf":[{"type":"integer","maximum":50.0,"minimum":1.0},{"type":"null"}],"title":"Task Count"},"id":{"type":"string","format":"uuid","title":"Id"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","user_id","created_at","updated_at"],"title":"UserPreferences"},"UserPreferencesUpsert":{"properties":{"productivity_period":{"anyOf":[{"type":"string","maxLength":40},{"type":"null"}],"title":"Productivity Period"},"work_style":{"anyOf":[{"type":"string","maxLength":80},{"type":"null"}],"title":"Work Style"},"focus_duration_minutes":{"anyOf":[{"type":"integer","maximum":180.0,"minimum":15.0},{"type":"null"}],"title":"Focus Duration Minutes"},"break_preference":{"anyOf":[{"type":"string","maxLength":80},{"type":"null"}],"title":"Break Preference"},"sleep_hours":{"anyOf":[{"type":"string","maxLength":80},{"type":"null"}],"title":"Sleep Hours"},"task_count":{"anyOf":[{"type":"integer","maximum":50.0,"minimum":1.0},{"type":"null"}],"title":"Task Count"},"routine_tasks":{"items":{"$ref":"#/components/schemas/RoutineTaskCreate"},"type":"array","title":"Routine Tasks"}},"type":"object","title":"UserPreferencesUpsert"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"Bearer":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}